% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_daisy_project.R
\name{create_daisy_project}
\alias{create_daisy_project}
\title{Create a new Daisy project directory structure}
\usage{
create_daisy_project(dir)
}
\arguments{
\item{dir}{The parent directory for your project}
}
\description{
Create a new Daisy project directory structure
}
