% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_daisy_scripts.R
\name{create_daisy_scripts}
\alias{create_daisy_scripts}
\title{Create daisy scripts using a dataframe and a template file}
\usage{
create_daisy_scripts(df, template_path, dir)
}
\arguments{
\item{df}{A dataframe containing the parameters for each simulation. Each row will correspond to one simulation script.}

\item{template_path}{The file path to a whisker template (.txt) that will be used to generate the daisy scripts. The template should use tags that match the column names in \code{df}.}

\item{dir}{The directory where the generated daisy scripts will be saved. Each simulation will be saved in a subfolder named after the \code{sim_name} column in \code{df}.}
}
\value{
Invisibly returns a character vector of the paths to the created .dai files.
}
\description{
Create daisy scripts using a dataframe and a template file
}
