% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_daisy_scripts.R
\name{run_daisy_scripts}
\alias{run_daisy_scripts}
\title{Run Daisy simulations in a specified directory, optionally in parallel.}
\usage{
run_daisy_scripts(
  dir,
  sims_to_run = "all",
  daisy_exe = "C:/Program Files/daisy 7.1.0/bin/daisy.exe",
  parallel = TRUE
)
}
\arguments{
\item{dir}{Directory containing Daisy simulation .dai files. The function will search for .dai files in all subdirectories. Note the output directory is defined in the daisy file (log_prefix) and must exist. Function assumes this is /outputs within each simulation folder and so creates this folder.}

\item{sims_to_run}{A character vector of simulation names to run, e.g.c("sim1", "sim2"). If "all", all .dai files found in the directory will be run.}

\item{daisy_exe}{The file path to the daisy executable. Defaults to "C:/Program Files/daisy 7.1.0/bin/daisy.exe". Update this if your daisy installation is in a different location.}

\item{parallel}{Whether to run simulations in parallel using available CPU cores. Defaults to TRUE.}
}
\value{
Invisibly returns a character vector of the .dai file paths that were run.
}
\description{
Run Daisy simulations in a specified directory, optionally in parallel.
}
