% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_daisy_outputs.R
\name{load_daisy_outputs}
\alias{load_daisy_outputs}
\title{Collate outputs from Daisy simulations into a list of data.tables}
\usage{
load_daisy_outputs(dir, required_outputs = "all")
}
\arguments{
\item{dir}{Directory containing the Daisy simulation outputs. The function will search for .dlf files in all subdirectories. Expected structure is dir/sim_name/outputsfile.dlf; sim_name is derived from two levels above each .dlf file.}

\item{required_outputs}{A character vector of output types to load, e.g. c("harvest", "crop_prod"). If "all", all unique .dlf file types will be loaded.}
}
\value{
A named list of data.tables, one per .dlf file type, with a sim column identifying the originating simulation.
}
\description{
Collate outputs from Daisy simulations into a list of data.tables
}
